emacs: rebind 'R' to meow-replace-save and bind '%' to meow-query-replace
I have rebound 'R' to meow-replace-save since I frequently need to
swap a piece of text in one part of the file with another piece of
text in another part of the file. I still want access to
meow-query-replace so I bound it to '%'. That isn't exactly the
easiest button to hit on standard qwerty keyboard but I also don't use
that command all that often.