emacs: remove automatically switching to meow-normal-mode in eshell
Everytime I executed an external command I would get put in normal
mode, even if it were a quick program like eza or ripgrep. This was
jaring and really broke to flow of running short cli
programs. However, it is the ideal behavior for long running
programs and programs with lots of output. Ideally, I would be dumped
into insert mode only if I hadn't run any cursor movement
commands. The next best thing I have to settle for is manually
defining which mode I want to be in. This isn't horrible since I've
integrated the char-modes pretty well but could be way better if
proper hooks were added for each of the modes and/or if a hook was
added for switching between modes (ideally the previous and next mode
would be passed as arguments to the function hooks).