emacs: move meow-grab to be under "'"
In the process, I have removed the embark bindings and I do not intend
to replace them. The bindings I have for them ("C-.") work fine and
just using that is way more consistent. I have also changed "<" and
">" to goto the begininning of the buffer and end of the buffer
respectively.