emacs: rename and reogranize cur-config modules
This regoranization and consolidation of my configuration modules is
intended to make naviating my config easier for me and whoever
stumbles across it. There was no good reason to have a module for
every single language, especially when most of it winds up depending
on what is defined in cur-config-ide.el anyways.