emacs: use emacs scroll up/down bindings instead of custom bindings
Scrolling up/down is something I find myself wanting to do outside of
normal state (both in motion state and in insert state). This makes it
so that using the standard emacs bindings for it (C-v and M-v) a good
idea. I have also removed the binding of RET to "C-c" as I want to use
the C-c bindings under the keypad state.