emacs: add isearch-backward-regexp to normal state
Unfortnunately when isearch-forward-regexp is given negative arg it
does not revesre the direction. Instead it changes from regexp to
normal isearch which is not what I want. This is a backup solution
which is to do it how vim and helix does. This leads to some jank
since its the only command where giving a negative arg doesn't reverse
the direction which kind of sucks but this works for now so its like
whatever.